Support for logger retrieval from Configuration
Added a get_logger() method to mandrel.config.Configuration,
which uses mandrel.bootstrap.get_logger in conjunction with
Configuration.NAME to get the right logger object.

Explicitly decided against storing logging objects in the
Configuration instances, as this is something that needs to
be sorted out in a case-by-case basis.