Add flag for controlling disabling of existing loggers
If you're starting a project with mandrel, you may move forward
confident in the idea that logging is configured once.

However, if you're adding it to an existing project, you may not
have such confidence.  In such cases, you may want to allow
the logging configuration to coexist with legacy uses of logging.

The mandrel.bootstrap.DISABLE_EXISTING_LOGGERS flag controls
how the "disable_existing_loggers" keyword param is handled when
applying the logging configuration file.  While it defaults to
True, consistent with logging.config.fileConfig's default,
you can override the flag in your bootstrap file and therefore
allow multiple logging configs to coexist.

It happens that this was necessary for the integration test
module I'm about to submit.