harness_loader for truly delightful dynamic loading frameworkyness.
Hark!  I want to put all my widgets in a particular location,
and load lots of them, but lazily to avoid import spaghetti.  And
I want to apply the same process to each at load time!

Waa laa:
    @harness_loader(convention_loader('all.mah.widgetz.%s_widget.Widget'))
    def widgety(widget_cls, name):
        return widget_cls(name)

    # Get an all.mah.widgetz.aard_widget.Widget
    # and dub it 'Thou Art Aard'.
    aard = widget('aard', 'Thou Art Aard')
    # And so on.
    vark = widget('vark', 'Varkiness')

I realize the documentation on this function sucks, but it's
not for lack of trying.